Fix fetching data from Yahoo! finance
Yahoo! finance recentrly became unaccessible, requiring cookies so the
calls can be made.

A recent "fix" was to revert to the v6 API, but even this was eventually
blocked by Yahoo!.

This change reverts to using the v7 API, and adds a way to fetch the
proper cookies and pass in the crumbs parameter in the API call. This
change was lifted and adapted from the one made in the `mop` project:
https://github.com/mop-tracker/mop/commit/c66e33ace3ff5d9c74f060830c9f04ae266d6e8f