Add examples for many of the possible calls
Add example for most of the types of calls that can be make with the
API. This was used to make sure the fixes made to support the cookie-
based calls are actually working when making several calls in a row.

Also remove broken to get the TWTR symbol as it's no longer traded and
resulted in errors being generated.